pull metadata implementation out into a crate
