+++ crates/wasm-metadata/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "wasm-metadata"
+version = "0.1.0"
+edition.workspace = true
+license = "Apache-2.0 WITH LLVM-exception"
+repository = "https://github.com/bytecodealliance/wasm-tools/tree/main/crates/wasm-metadata"
+description = "Read and manipulate WebAssembly metadata"
+
+[dependencies]
+clap = { workspace = true, optional = true }
+anyhow = { workspace = true }
+wasmparser = { workspace = true }
+wasm-encoder = { workspace = true }
+indexmap = { workspace = true }
